requirements.txt
requests
django
djangorestframework
mysqlclient
django-cors-headers
python-decouple
djangorestframework-simplejwt
drf-yasg
django_redis